Create requirements.txt
These are the additional packages installed to run the application that are not already prepackaged in the Treehouse workspace console.  Adding this package should allow the app to run fine.  I built the app in workspaces instead of a vitrualenv, so it doesn't populate all the python  installed packages that are standard within Workspaces only the additional ones I installed.